requirements.txt
backoff==2.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.13
Flask==2.2.3
googleapis-common-protos==1.58.0
grpcio==1.51.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
opentelemetry-api==1.16.0
opentelemetry-distro==0.37b0
opentelemetry-exporter-otlp==1.16.0
opentelemetry-exporter-otlp-proto-grpc==1.16.0
opentelemetry-exporter-otlp-proto-http==1.16.0
opentelemetry-instrumentation==0.37b0
opentelemetry-instrumentation-aws-lambda==0.37b0
opentelemetry-instrumentation-dbapi==0.37b0
opentelemetry-instrumentation-flask==0.37b0
opentelemetry-instrumentation-jinja2==0.37b0
opentelemetry-instrumentation-logging==0.37b0
opentelemetry-instrumentation-sqlite3==0.37b0
opentelemetry-instrumentation-urllib==0.37b0
opentelemetry-instrumentation-wsgi==0.37b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-semantic-conventions==0.37b0
opentelemetry-util-http==0.37b0
protobuf==4.22.1
pyaml==21.10.1
PyYAML==6.0
requests==2.28.2
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
